backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# AI and Embeddings
openai==1.10.0
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
sentence-transformers==2.3.1
tiktoken==0.5.2

# Vector Database
qdrant-client==1.7.3

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
Pillow==10.2.0
pytesseract==0.3.10
unstructured==0.12.0

# Google Drive Integration
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.115.0

# Email
fastapi-mail==1.4.1

# Utilities
httpx==0.26.0
aiofiles==23.2.1
redis==5.0.1
celery==5.3.6
python-slugify==8.0.1
